# docs/swagger.yaml
definitions:
...
  entity.UserRole:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      id:
        type: string
      role_id:
        type: string
      updatedAt:
        type: string
      updatedBy:
        type: string
      user_id:
        type: string
    type: object
  helper.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
...
  request.UserHasRole:
    properties:
      roles:
        items:
          $ref: '#/definitions/entity.UserRole'
        type: array
    type: object
  request.UserUpdateRequest:
    properties:
      created_by:
        description: Password  string `json:"password" validate:"required,min=8"`
        type: string
      email:
        type: string
      name:
        type: string
      updated_by:
        type: string
      username:
        type: string
    required:
    - email
    - name
    - username
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Api Service boilerplate dari golang gin dan gorm
  termsOfService: http://swagger.io/terms/
  title: My Admin Service Api
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: REST API for Users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerAuth: []
      summary: Get list Users
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: REST API for Users
      parameters:
      - description: Users ID
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerAuth: []
      summary: Delete Users
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: REST API for Users
      parameters:
      - description: Users ID
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerAuth: []
      summary: Get one Users
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: REST API for Users
      parameters:
      - description: Users ID
        in: query
        name: id
        type: string
      - description: Input data in JSON format
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerAuth: []
      summary: Update Users
      tags:
      - Users
  /users/roles/assign:
    post:
      consumes:
      - application/json
      description: REST API User Role
      parameters:
      - description: Input data in JSON format
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.UserHasRole'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerAuth: []
      summary: Attach user has role
      tags:
      - Users
  /users/roles/unassign:
    delete:
      consumes:
      - application/json
      description: REST API User Role
      parameters:
      - description: Users ID
        in: query
        name: id
        type: string
      - description: Input data in JSON format
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.UserHasRole'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerAuth: []
      summary: Detach user has role
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"